+++ src/main/java/wooteco/subway/domain/member/Member.java
@@ -4,7 +4,7 @@ import org.apache.commons.lang3.StringUtils;
 import org.springframework.data.annotation.Id;
 import org.springframework.data.relational.core.mapping.Embedded;
 import wooteco.subway.exception.DuplicatedFavoriteException;
-import wooteco.subway.exception.InvalidAuthenticationException;
+import wooteco.subway.exception.authentication.InvalidAuthenticationException;
 
 public class Member {
     @Id
@@ -62,7 +62,7 @@ public class Member {
 
     public void validateId(Long id) {
         if (!this.id.equals(id)) {
-            throw new InvalidAuthenticationException("니 아이디 아님");
+            throw new InvalidAuthenticationException("니 아이디 아니야");
         }
     }
