+++ src/main/java/wooteco/subway/exception/authentication/InvalidSessionException.java
@@ -1,0 +1,7 @@
+package wooteco.subway.exception.authentication;
+
+public class InvalidSessionException extends InvalidAuthenticationException{
+    public InvalidSessionException() {
+        super("세션이 이상해요");
+    }
+}
