+++ src/main/java/wooteco/subway/exception/authentication/InvalidTokenException.java
@@ -1,0 +1,7 @@
+package wooteco.subway.exception.authentication;
+
+public class InvalidTokenException extends InvalidAuthenticationException{
+    public InvalidTokenException() {
+        super("토큰이 이상해요");
+    }
+}
