+++ src/main/java/wooteco/subway/exception/notexist/NoResourceExistException.java
@@ -1,4 +1,4 @@
-package wooteco.subway.exception;
+package wooteco.subway.exception.notexist;
 
 public class NoResourceExistException extends IllegalArgumentException {
     public NoResourceExistException(String message) {
