+++ src/main/java/wooteco/subway/exception/notexist/NoStationExistsException.java
@@ -1,4 +1,4 @@
-package wooteco.subway.exception;
+package wooteco.subway.exception.notexist;
 
 public class NoStationExistsException extends NoResourceExistException {
     public NoStationExistsException() {
