+++ src/main/java/wooteco/subway/service/line/LineStationService.java
@@ -6,7 +6,7 @@ import wooteco.subway.domain.line.LineRepository;
 import wooteco.subway.domain.line.Lines;
 import wooteco.subway.domain.station.Station;
 import wooteco.subway.domain.station.StationRepository;
-import wooteco.subway.exception.NoLineExistException;
+import wooteco.subway.exception.notexist.NoLineExistException;
 import wooteco.subway.service.line.dto.LineDetailResponse;
 import wooteco.subway.service.line.dto.WholeSubwayResponse;
