+++ src/main/java/wooteco/subway/service/member/MemberService.java
@@ -4,8 +4,9 @@ import org.springframework.stereotype.Service;
 import wooteco.subway.domain.member.Member;
 import wooteco.subway.domain.member.MemberRepository;
 import wooteco.subway.exception.DuplicatedEmailException;
-import wooteco.subway.exception.InvalidAuthenticationException;
-import wooteco.subway.exception.NoMemberExistException;
+import wooteco.subway.exception.authentication.InvalidPasswordException;
+import wooteco.subway.exception.authentication.InvalidSessionException;
+import wooteco.subway.exception.notexist.NoMemberExistException;
 import wooteco.subway.infra.JwtTokenProvider;
 import wooteco.subway.service.member.dto.LoginRequest;
 import wooteco.subway.service.member.dto.UpdateMemberRequest;
@@ -40,7 +41,7 @@ public class MemberService {
     public String createToken(LoginRequest param) {
         Member member = memberRepository.findByEmail(param.getEmail()).orElseThrow(NoMemberExistException::new);
         if (!member.checkPassword(param.getPassword())) {
-            throw new InvalidAuthenticationException("잘못된 패스워드");
+            throw new InvalidPasswordException();
         }
 
         return jwtTokenProvider.createToken(param.getEmail());
@@ -53,7 +54,7 @@ public class MemberService {
     public Member loginWithForm(String email, String password) {
         Member member = findMemberByEmail(email);
         if (!member.checkPassword(password)) {
-            throw new InvalidAuthenticationException("비번 틀렸어");
+            throw new InvalidSessionException();
         }
         return member;
     }
