+++ src/main/java/wooteco/subway/web/member/interceptor/SessionInterceptor.java
@@ -3,7 +3,7 @@ package wooteco.subway.web.member.interceptor;
 import org.apache.logging.log4j.util.Strings;
 import org.springframework.stereotype.Component;
 import org.springframework.web.servlet.HandlerInterceptor;
-import wooteco.subway.exception.InvalidAuthenticationException;
+import wooteco.subway.exception.authentication.InvalidSessionException;
 
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
@@ -19,7 +19,7 @@ public class SessionInterceptor implements HandlerInterceptor {
 
         String email = (String) request.getSession().getAttribute("loginMemberEmail");
         if (Strings.isBlank(email)) {
-            throw new InvalidAuthenticationException("Session이 잘못되었어요.");
+            throw new InvalidSessionException();
         }
 
         request.setAttribute("loginMemberEmailSession", email);
